Refactoring Attribute to NodeAttribute
Upon further reflection, a system has a set of attributes that will be
assigned to each node; Thus we need both an Attribute, which provides
definition, and a NodeAttribute that connects a Node to an Attribute.
There is knowledge of Node implied in the name, which will probably be
factored again.